+++ src/Model/Field.php
@@ -126,6 +126,21 @@ class Field extends DataObject
 		return $fields;
 	}
 
+	public function getFieldExportData()
+	{
+		$data = [];
+		foreach($this->toMap() as $fieldName => $fieldValue)
+		{
+			if (!preg_match('/ID$/', $fieldName))
+			{
+				$data[$fieldName] = $fieldValue;
+			}
+		}
+		unset($data['RecordClassName'], $data['LastEdited'], $data['Created'], $data['ContainerClass']);
+		$this->invokeWithExtensions('updateFieldExportData', $data);
+		return $data;
+	}
+
 	public function onBeforeDuplicate($original, $doWrite, $relations)
 	{
 		$this->Name = $name = 'Copy of '.$original->Name;
@@ -257,7 +272,7 @@ class Field extends DataObject
 
 	public function Explain()
 	{
-		$text = '<div><strong>'.$this->Name.'</strong><div><div>'.$this->singular_name().'</div>';
+		$text = '<div data-field-name><strong>'.$this->Name.'</strong></div><div>'.$this->singular_name().'</div>';
 		if ($this->HideByDefault)
 		{
 			$text .= '<em>(Default Hidden)</em>';